Move metadata to @desmodder/common
