+++ package.json
@@ -16,7 +16,7 @@
     "#plugins/*": "./src/plugins/*",
     "#parsing/*": "./common/parsing/*",
     "#tests": "./src/tests/puppeteer-utils.ts",
-    "#metadata/*": "./metadata/*",
+    "#metadata/*": "./common/metadata/*",
     "#text-mode-core": "./text-mode-core/index.ts",
     "#text-mode-editor": "./text-mode-editor/index.ts",
     "#components": "./common/globals/components/index.ts"
