+++ src/tsconfig.json
@@ -4,8 +4,7 @@
   "references": [
     { "path": "../text-mode-core" },
     { "path": "../localization" },
-    { "path": "../common" },
-    { "path": "../metadata" }
+    { "path": "../common" }
   ],
   "include": ["**/*"]
 }
